Got D-Viz Building on Linux
+ It took a little bit of work, but I got D-Viz running on Mint 18.1 Cinnamon 64-bit, using GCC 7.1.
+ Moving to a top-shelf compiler exposed a few issues that MSVC was willing to let slide. Some of those issues were exposed in my own submodules, and I'll have to submit those fixes separately.
+ A few features have yet to be ported over. The main one is Xbox controller support.

+++ source/Scene/visualizationAsset.cpp
@@ -223,7 +223,7 @@ bool VisualizationAsset::InitializeBlockTransformations()
       /* type = */ GL_FLOAT,
       /* normalized = */ GL_FALSE,
       /* stride = */ sizeOfMatrix,
-      /* ptr = */ static_cast<GLvoid*>(0 * sizeOfVector));
+      /* ptr = */ reinterpret_cast<GLvoid*>(0 * sizeOfVector));
 
    // Row 2 of the matrix:
    m_graphicsDevice.glEnableVertexAttribArray(2);
@@ -314,7 +314,7 @@ std::uint32_t VisualizationAsset::LoadBufferData(
    FindLargestDirectory(tree);
 
    assert(m_blockColors.size() == m_blockTransformations.size());
-   assert(m_blockColors.size() == m_blockCount);
+   assert(m_blockColors.size() == static_cast<int>(m_blockCount));
 
    return m_blockCount;
 }
